Create ImmersiveSizeFunction
Hi, i downloaded your extension source, and it was really usefull to me. Congratulations!
To thank, i propose the use of just one function to get the size instead of using two (one for width and another for height).
If you need any help with BlueTooth or android Volume, feel free to contact me.
Thank you!